Title: 달에서 무게 구하기, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240306/달에서 무게 구하기/weight-on-the-moon.java
@@ -2,8 +2,8 @@ public class Main {
     public static void main(String[] args) {
         // 여기에 코드를 작성해주세요.
         int a;
-        float b;
-        float c;
+        double b;
+        double c;
         a = 13;
         b = 0.165;
         c = a*b;
